# 02_geo_service_with_swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "You can get address suggestion by address strind.",
        "title": "Geo Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/address/geocode": {
            "post": {
                "description": "get address suggestions from location",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geocode"
                ],
                "summary": "GeoCode",
                "operationId": "geocode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "latitude",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "longitude",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Address"
                            }
                        }
                    }
                }
            }
        },
        "/api/address/search": {
            "post": {
                "description": "get address suggestions from address string",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "AddressSearcb",
                "operationId": "search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address string",
                        "name": "input",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Address"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "house": {
                    "type": "string"
                },
                "lat": {
                    "type": "string"
                },
                "lon": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        }
    }
}